requirements.txt
numpy==1.23.5
pandas==1.5.2
protobuf==4.21.11
python-dotenv==1.0.0
scikit_learn==1.1.3
splitwise==2.5.0
google-api-core==2.10.1
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-bigquery==3.4.1
google-cloud-bigquery-storage==2.11.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
pyarrow==10.0.1